tests/wpt/web-platform-tests/html/canvas/offscreen/the-offscreen-canvas/size.large.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: size.large</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>size.large</h1>
<p class="desc"></p>

<p class="notes">Not sure how reasonable this is, but the spec doesn't say there's an upper limit on the size.
<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var n = 2147483647; // 2^31 - 1, which should be supported by any sensible definition of "long"
offscreenCanvas.width = n;
offscreenCanvas.height = n;
_assertSame(offscreenCanvas.width, n, "offscreenCanvas.width", "n");
_assertSame(offscreenCanvas.height, n, "offscreenCanvas.height", "n");
t.done();

});
</script>
